Refactor call adapter and factories
