+++ sandwich/src/main/kotlin/com/skydoves/sandwich/adapters/DataSourceCallAdapter.kt
@@ -14,8 +14,10 @@
  * limitations under the License.
  */
 
-package com.skydoves.sandwich
+package com.skydoves.sandwich.adapters
 
+import com.skydoves.sandwich.DataSource
+import com.skydoves.sandwich.ResponseDataSource
 import retrofit2.Call
 import retrofit2.CallAdapter
 import java.lang.reflect.Type
