+++ sandwich/src/main/kotlin/com/skydoves/sandwich/adapters/DataSourceCallAdapterFactory.kt
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package com.skydoves.sandwich
+package com.skydoves.sandwich.adapters
 
+import com.skydoves.sandwich.DataSource
 import retrofit2.CallAdapter
 import retrofit2.Retrofit
 import java.lang.reflect.ParameterizedType
