+++ sandwich/src/main/kotlin/com/skydoves/sandwich/coroutines/CoroutinesDataSourceCallAdapter.kt
@@ -29,7 +29,7 @@ import java.lang.reflect.Type
  *
  * request API network call asynchronously and returns [DataSource].
  */
-public class CoroutinesDataSourceCallAdapter constructor(
+internal class CoroutinesDataSourceCallAdapter constructor(
   private val responseType: Type
 ) : CallAdapter<Type, Call<DataSource<Type>>> {
