+++ sandwich/src/main/kotlin/com/skydoves/sandwich/coroutines/CoroutinesDataSourceCallAdapterFactory.kt
@@ -41,7 +41,7 @@ public class CoroutinesDataSourceCallAdapterFactory private constructor() : Call
     returnType: Type,
     annotations: Array<Annotation>,
     retrofit: Retrofit
-  ): CoroutinesDataSourceCallAdapter? = when (getRawType(returnType)) {
+  ): CallAdapter<*, *>? = when (getRawType(returnType)) {
     Call::class.java -> {
       val callType = getParameterUpperBound(0, returnType as ParameterizedType)
       when (getRawType(callType)) {
